requirements.txt
# ========================================================
# Archivo de Requerimientos para Asistente Legal Multimodal
# Versión final con pysqlite3-binary CORRECTO para compatibilidad en Render
# ========================================================

# --- Framework Principal ---
fastapi==0.111.0
uvicorn==0.29.0
sqlmodel==0.0.18
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-multipart==0.0.9

# --- Contraseñas y Seguridad ---
passlib==1.7.4
argon2-cffi==23.1.0
python-jose[cryptography]==3.3.0

# --- Tareas Asíncronas ---
# celery==5.4.0
# redis==5.0.4

# --- Ecosistema de IA ---
langchain==0.2.5
langchain-core==0.2.8
langchain-google-genai==1.0.6
google-generativeai==0.5.4
langgraph==0.1.1
chromadb==0.4.24
pypdf==4.2.0
# =====================================================================
# CORRECCIÓN CLAVE: Usamos una versión de pysqlite3-binary que SÍ existe.
# =====================================================================
pysqlite3-binary==0.5.4

# --- Utilidades ---
python-docx==1.1.2
docx2txt==0.8

# --- Generación de Documentos ---
fpdf2==2.7.8

# --- Comunicaciones ---
resend==0.8.0


# SpeechRecognition==3.10.0
# pydub==0.25.1
# # ffmpeg-python==0.2.0

#unidecode # Para normalización de texto (opcional)
#pandas openpyxl # Para manejo avanzado de datos y Excel (opcional)


#ragas # Para recuperación avanzada de información (opcional)
#datasets # Para manejo de conjuntos de datos (opcional)
langchain-chroma
